FE: Added login and registration using localstorage for verify endpoints by bearer token headers

+++ frontend/src/app/services/class/classService.tsx
@@ -2,15 +2,18 @@ export const callGetAllClasses = async (
   bucketName: string,
   databaseName: string
 ) => {
+  const token = localStorage.getItem("access_token");
+  if (!token) {
+    throw new Error("User is not authenticated");
+  }
+
   const response = await fetch(
-    "http://localhost:8000/classes/get-all-classes/" +
-      bucketName +
-      "/" +
-      databaseName,
+    `http://localhost:8000/classes/get-all-classes/${bucketName}/${databaseName}`,
     {
       method: "GET",
       headers: {
         "Content-Type": "application/json",
+        Authorization: `Bearer ${token}`,
       },
     }
   );
@@ -27,17 +30,18 @@ export const callRemoveClass = async (
   databaseName: string,
   className: string
 ) => {
+  const token = localStorage.getItem("access_token");
+  if (!token) {
+    throw new Error("User is not authenticated");
+  }
+
   const response = await fetch(
-    "http://localhost:8000/classes/delete-class/" +
-      bucketName +
-      "/" +
-      databaseName +
-      "/" +
-      className,
+    `http://localhost:8000/classes/delete-class/${bucketName}/${databaseName}/${className}`,
     {
       method: "DELETE",
       headers: {
         "Content-Type": "application/json",
+        Authorization: `Bearer ${token}`,
       },
     }
   );
