+++ frontend/src/app/services/dataset/datasetService.tsx
@@ -4,6 +4,11 @@ export const callUploadZip = async (
   expectedNumFilesPerClass: number,
   file: File
 ) => {
+  const token = localStorage.getItem("access_token");
+  if (!token) {
+    throw new Error("User is not authenticated");
+  }
+
   const formData = new FormData();
   formData.append("file", file);
 
@@ -12,6 +17,9 @@ export const callUploadZip = async (
   const response = await fetch(url, {
     method: "POST",
     body: formData,
+    headers: {
+      Authorization: `Bearer ${token}`,
+    },
   });
 
   if (!response.ok) {
@@ -22,12 +30,18 @@ export const callUploadZip = async (
 };
 
 export const callGetAllDatasets = async (bucketName: string) => {
+  const token = localStorage.getItem("access_token");
+  if (!token) {
+    throw new Error("User is not authenticated");
+  }
+
   const response = await fetch(
-    "http://localhost:8000/databases/get-all-databases/" + bucketName,
+    `http://localhost:8000/databases/get-all-databases/${bucketName}`,
     {
       method: "GET",
       headers: {
         "Content-Type": "application/json",
+        Authorization: `Bearer ${token}`,
       },
     }
   );
@@ -43,15 +57,18 @@ export const callRemoveDataset = async (
   bucketName: string,
   databaseName: string
 ) => {
+  const token = localStorage.getItem("access_token");
+  if (!token) {
+    throw new Error("User is not authenticated");
+  }
+
   const response = await fetch(
-    "http://localhost:8000/databases/delete-database/" +
-      bucketName +
-      "/" +
-      databaseName,
+    `http://localhost:8000/databases/delete-database/${bucketName}/${databaseName}`,
     {
       method: "DELETE",
       headers: {
         "Content-Type": "application/json",
+        Authorization: `Bearer ${token}`,
       },
     }
   );
