+++ frontend/src/app/services/sample/sampleService.tsx
@@ -3,17 +3,18 @@ export const callGetAllSamples = async (
   databaseName: string,
   className: string
 ) => {
+  const token = localStorage.getItem("access_token");
+  if (!token) {
+    throw new Error("User is not authenticated");
+  }
+
   const response = await fetch(
-    "http://localhost:8000/samples/get-all-samples-in-class/" +
-      bucketName +
-      "/" +
-      databaseName +
-      "/" +
-      className,
+    `http://localhost:8000/samples/get-all-samples-in-class/${bucketName}/${databaseName}/${className}`,
     {
       method: "GET",
       headers: {
         "Content-Type": "application/json",
+        Authorization: `Bearer ${token}`,
       },
     }
   );
@@ -25,25 +26,24 @@ export const callGetAllSamples = async (
   return await response.json();
 };
 
-export const callRemoveDataset = async (
+export const callRemoveSample = async (
   bucketName: string,
   databaseName: string,
   className: string,
   sampleName: string
 ) => {
+  const token = localStorage.getItem("access_token");
+  if (!token) {
+    throw new Error("User is not authenticated");
+  }
+
   const response = await fetch(
-    "http://localhost:8000/samples/delete-sample/" +
-      bucketName +
-      "/" +
-      databaseName +
-      "/" +
-      className +
-      "/" +
-      sampleName,
+    `http://localhost:8000/samples/delete-sample/${bucketName}/${databaseName}/${className}/${sampleName}`,
     {
       method: "DELETE",
       headers: {
         "Content-Type": "application/json",
+        Authorization: `Bearer ${token}`,
       },
     }
   );
